+++ ereuse_devicehub/resources/action/models.py
@@ -1773,6 +1773,14 @@ class MoveOnDocument(JoinedTableMixin, ActionWithMultipleTradeDocuments):
     container_to_id.comment = """This is the trade document used as container in a outgoing lot"""
 
 
+class Delete(ActionWithMultipleDevices):
+    # TODO in a new architecture we need rename this class to Deactivate
+
+    """The act save in device who and why this devices was delete.
+    We never delete one device, but we can deactivate."""
+    pass
+
+
 class Migrate(JoinedTableMixin, ActionWithMultipleDevices):
     """Moves the devices to a new database/inventory. Devices cannot be
     modified anymore at the previous database.