+++ ereuse_devicehub/resources/device/models.py
@@ -127,6 +127,7 @@ class Device(Thing):
     allocated.comment = "device is allocated or not."
     devicehub_id = db.Column(db.CIText(), nullable=True, unique=True, default=create_code)
     devicehub_id.comment = "device have a unique code."
+    active = db.Column(Boolean, default=True)
 
     _NON_PHYSICAL_PROPS = {
         'id',
@@ -150,7 +151,8 @@ class Device(Thing):
         'sku',
         'image',
         'allocated',
-        'devicehub_id'
+        'devicehub_id',
+        'active'
     }
 
     __table_args__ = (