+++ tests/test_action.py
@@ -2822,6 +2822,75 @@ def test_moveOnDocument(user: UserClient, user2: UserClient):
 
 @pytest.mark.mvp
 @pytest.mark.usefixtures(conftest.app_context.__name__)
+def test_delete_devices(user: UserClient):
+    """This action deactive one device and simulate than one devices is delete."""
+
+    snap, _ = user.post(file('acer.happy.battery.snapshot'), res=models.Snapshot)
+    request = {'type': 'Delete', 'devices': [snap['device']['id']], 'name': 'borrado universal', 'severity': 'Info', 'description': 'duplicity of devices', 'endTime': '2021-07-07T22:00:00.000Z'}
+
+    action, _ = user.post(res=models.Action, data=request)
+
+    # Check get one device
+    user.get(res=Device, item=snap['device']['devicehubID'], status=404)
+    db_device = Device.query.filter_by(id=snap['device']['id']).one()
+
+    action_delete = sorted(db_device.actions, key=lambda x: x.created)[-1]
+
+    assert action_delete.t == 'Delete'
+    assert str(action_delete.id) == action['id']
+    assert db_device.active == False
+
+
+
+    # Check use of filter from frontend
+    url = '/devices/?filter={"type":["Computer"]}'
+
+    devices, res = user.get(url, None)
+    assert len(devices['items']) == 0
+
+
+@pytest.mark.mvp
+@pytest.mark.usefixtures(conftest.app_context.__name__)
+def test_delete_devices_check_sync(user: UserClient):
+    """This action deactive one device and simulate than one devices is delete."""
+
+    file_snap1 = file('1-device-with-components.snapshot')
+    file_snap2 = file('2-device-with-components.snapshot')
+    snap, _ = user.post(file_snap1, res=models.Snapshot)
+    request = {'type': 'Delete', 'devices': [snap['device']['id']], 'name': 'borrado universal', 'severity': 'Info', 'description': 'duplicity of devices', 'endTime': '2021-07-07T22:00:00.000Z'}
+
+    action, _ = user.post(res=models.Action, data=request)
+
+    device1 = Device.query.filter_by(id=snap['device']['id']).one()
+
+    snap2, _ = user.post(file_snap2, res=models.Snapshot)
+    request2 = {'type': 'Delete', 'devices': [snap2['device']['id']], 'name': 'borrado universal', 'severity': 'Info', 'description': 'duplicity of devices', 'endTime': '2021-07-07T22:00:00.000Z'}
+
+    action2, _ = user.post(res=models.Action, data=request2)
+
+    device2 = Device.query.filter_by(id=snap2['device']['id']).one()
+    # check than device2 is an other device than device1
+    assert device2.id != device1.id
+    # check than device2 have the components of device1
+    assert len([x for x in device2.components
+        if device1.id in [y.device.id for y in x.actions if hasattr(y, 'device')]]) == 1
+
+
+@pytest.mark.mvp
+@pytest.mark.usefixtures(conftest.app_context.__name__)
+def test_delete_devices_permitions(user: UserClient, user2: UserClient):
+    """This action deactive one device and simulate than one devices is delete."""
+
+    file_snap = file('1-device-with-components.snapshot')
+    snap, _ = user.post(file_snap, res=models.Snapshot)
+    device = Device.query.filter_by(id=snap['device']['id']).one()
+
+    request = {'type': 'Delete', 'devices': [snap['device']['id']], 'name': 'borrado universal', 'severity': 'Info', 'description': 'duplicity of devices', 'endTime': '2021-07-07T22:00:00.000Z'}
+    action, _ = user2.post(res=models.Action, data=request, status=422)
+
+
+@pytest.mark.mvp
+@pytest.mark.usefixtures(conftest.app_context.__name__)
 def test_moveOnDocument_bug168(user: UserClient, user2: UserClient):
     """If you use one moveOnDocument in a trade Document. Next you can not drop this document."""
     lotIn, _ = user.post({'name': 'MyLotIn'}, res=Lot)